Add files via upload
A/B Testing
A/B Testing means analyzing two marketing strategies to choose the best marketing strategy that can convert more traffic into sales (or more traffic into your desired goal) effectively and efficiently. A/B testing is one of the valuable concepts that every Data Science professional should know. In this article, I will take you through the task of A/B Testing using Python.

In A/B testing, we analyze the results of two marketing strategies to choose the best one for future marketing campaigns. For example, when I started an ad campaign on Instagram to promote my Instagram post for the very first time, my target audience was different from the target audience of my second ad campaign. After analyzing the results of both ad campaigns, I always preferred the audience of the second ad campaign as it gave better reach and followers than the first one.

That is what A/B testing means. Your goal can be to boost sales, followers, or traffic, but when we choose the best marketing strategy according to the results of our previous marketing campaigns, it is nothing but A/B testing.

For the task of A/B testing using Python, we need to have a dataset about two different marketing strategies for the same goal. I found a dataset that will help us perform A/B testing. You can download it from (https://statso.io/a-b-testing-case-study/).

In the section below, I will take you through the task of A/B Testing using Python.

A/B Testing Case Study
A/B testing helps in finding a better approach to finding customers, marketing products, getting a higher reach, or anything that helps a business convert most of its target customers into actual customers.

Campaign Name: The name of the campaign Date: Date of the record Spend: Amount spent on the campaign in dollars of Impressions: Number of impressions the ad crossed through the campaign Reach: The number of unique impressions received in the ad of Website Clicks: Number of website clicks received through the ads of Searches: Number of users who performed searches on the website of View Content: Number of users who viewed content and products on the website of Add to Cart: Number of users who added products to the cart of Purchase: Number of purchases Two campaigns were performed by the company:

Control Campaign Test Campaign Perform A/B testing to find the best campaign for the company to get more customers.